Generate/Fetch Databricks Host

If both id and prefix are NULL function will search for the DATABRICKS_HOST environment variable.

When defining id and prefix you do not need to specify the whole URL. E.g. https://<prefix>.<id>.cloud.databricks.com/ is the form to follow.

Usage

db_host(id = NULL, prefix = NULL, profile = getOption("db_profile", NULL))

Arguments

id

The workspace string

prefix

Workspace prefix

profile

Profile to use when fetching from environment variable (e.g. .Renviron) or .databricksfg file

Value

workspace URL

Details

The behaviour is subject to change depending if db_profile and use_databrickscfg options are set.

use_databrickscfg: Boolean (default: FALSE), determines if credentials are fetched from profile of .databrickscfg or .Renviron

db_profile: String (default: NULL), determines profile used.

See vignette on authentication for more details.

See also