Create SQL Endpoint

Create SQL Endpoint

Usage

db_sql_endpoint_create(
  name,
  cluster_size,
  min_num_clusters = 1,
  max_num_clusters = 1,
  auto_stop_mins = 30,
  tags = list(),
  spot_instance_policy = c("COST_OPTIMIZED", "RELIABILITY_OPTIMIZED"),
  enable_photon = TRUE,
  enable_serverless_compute = NULL,
  channel = c("CHANNEL_NAME_CURRENT", "CHANNEL_NAME_PREVIEW"),
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)

Arguments

name

Name of the SQL endpoint. Must be unique.

cluster_size

Size of the clusters allocated to the endpoint. One of 2X-Small, X-Small, Small, Medium, Large, X-Large, 2X-Large, 3X-Large, 4X-Large.

min_num_clusters

Minimum number of clusters available when a SQL endpoint is running. The default is 1.

max_num_clusters

Maximum number of clusters available when a SQL endpoint is running. If multi-cluster load balancing is not enabled, this is limited to 1.

auto_stop_mins

Time in minutes until an idle SQL endpoint terminates all clusters and stops. Defaults to 30. For Serverless SQL endpoints (enable_serverless_compute = TRUE), set this to 10.

tags

Named list that describes the endpoint. Databricks tags all endpoint resources with these tags.

spot_instance_policy

The spot policy to use for allocating instances to clusters. This field is not used if the SQL endpoint is a Serverless SQL endpoint.

enable_photon

Whether queries are executed on a native vectorized engine that speeds up query execution. The default is TRUE.

enable_serverless_compute

Whether this SQL endpoint is a Serverless endpoint. To use a Serverless SQL endpoint, you must enable Serverless SQL endpoints for the workspace. If Serverless SQL endpoints are disabled for the workspace, the default is FALSE If Serverless SQL endpoints are enabled for the workspace, the default is TRUE.

channel

Whether to use the current SQL endpoint compute version or the preview version. Databricks does not recommend using preview versions for production workloads. The default is CHANNEL_NAME_CURRENT.

host

Databricks workspace URL, defaults to calling db_host().

token

Databricks workspace token, defaults to calling db_token().

perform_request

If TRUE (default) the request is performed, if FALSE the httr2 request is returned without being performed.

See also